# backend/application/tags/views.py
from typing import Any

from fastapi import APIRouter, Depends
from fastapi.responses import JSONResponse
from starlette.status import HTTP_200_OK, HTTP_404_NOT_FOUND

from application.auth.permissions import IsAdmin, IsAuthenticated, PermissionsDependency
from application.managers import Manager
from application.schemas import SearchName
from application.settings import NOT_FOUND
from application.tags.models import Tag
from application.tags.schemas import TagCreate, TagOut, TagUpdate

router = APIRouter()
admin_router = APIRouter()
tag = Manager(Tag)


@admin_router.post("/", status_code=HTTP_200_OK)
async def create_tag(tag_in: TagCreate) -> JSONResponse:
    """Создавать теги может только администратор."""
    return await tag.create(tag_in.dict())


@router.get("/", response_model=list[TagOut], status_code=HTTP_200_OK)
async def get_tags(params: SearchName = Depends()) -> dict:
    """Cписок тегов."""
    return await tag.get_all_list(
        params,
        attr_name="name",
        query=Tag.list_columns("id", "name", "color", "slug"),
    )


@router.get("/{tag_id}/", response_model=TagOut, status_code=HTTP_200_OK)
async def details_tag(tag_id: int) -> Any:
    """Получение тега."""
    return await tag.by_id(pk=tag_id) or NOT_FOUND


@admin_router.patch("/{tag_id}/", response_model=TagOut, status_code=HTTP_200_OK)
async def update_tag(tag_id: int, items: TagUpdate) -> JSONResponse:
    """Редактировать теги может только администратор."""
    items = {k: v for k, v in items.dict().items() if v}
    return await tag.update(items=items, pk=tag_id)


@admin_router.delete("/{tag_id}/", status_code=HTTP_200_OK)
async def delete_tag(tag_id: int) -> JSONResponse:
    if not await tag.delete(pk=tag_id):
        return JSONResponse({"detail": "Тег не существует"}, HTTP_404_NOT_FOUND)


router.include_router(
    admin_router,
    dependencies=[Depends(PermissionsDependency([IsAuthenticated, IsAdmin]))],
)